requirements.txt
pytest==3.9.3
PyOpenGL==3.1.0
PySDL2==0.9.6
